requirements.txt
langchain
langchain_ollama
langchain_core
python-dotenv
rich
pydantic


